Move out vg join
